Use emacs keystrokes
This commit will change the default editing style from vim to emacs.

I personally prefer the emacs editing style compared to vim's after
using emacs for a couple of months.

On a personal note, I am just changing the style as the first change as
I think I will find it tough to work with any files in the vim bindings.