Add support for LSP with eglot
I recently felt that a need for a consistent "interface" to interact
with certain language specific features was needed. However, I did not
want to use "bloat" which is what lsp-mode.el feels like to me now.

So, I downloaded eglot, which is perfect! This commit also adds
support for lsp for python and ruby modes.

Note: You need to download the language servers yourself. I use `pylsp`
for python and `solargraph` for ruby.